requirements.txt
anthropic==0.18.1
streamlit==1.32.0
pandas==2.2.0
python-dotenv==1.0.1
numpy==1.26.4
scikit-learn==1.4.1.post1
plotly==5.18.0
biopython==1.83 